requirements.txt
openai==0.28.1
python-docx  
streamlit